MAILBOX-314 add a Integration test for search with ES without TIKA
This is for speed reason
