+++ mailbox/elasticsearch/src/test/java/org/apache/james/mailbox/elasticsearch/ElasticSearchV1IntegrationTestWithoutTika.java
@@ -19,6 +19,7 @@
 
 package org.apache.james.mailbox.elasticsearch;
 
+import java.net.URISyntaxException;
 import java.time.ZoneId;
 
 import org.apache.james.mailbox.elasticsearch.json.MessageToElasticSearchJson;
@@ -30,11 +31,40 @@ import org.apache.james.mailbox.elasticsearch.search.ElasticSearchSearcherV1;
 import org.apache.james.mailbox.extractor.TextExtractor;
 import org.apache.james.mailbox.model.MailboxId;
 import org.apache.james.mailbox.model.MessageId;
+import org.apache.james.mailbox.store.extractor.DefaultTextExtractor;
+import org.apache.james.mailbox.tika.TikaConfiguration;
+import org.apache.james.mailbox.tika.TikaContainer;
+import org.apache.james.mailbox.tika.TikaHttpClientImpl;
+import org.apache.james.mailbox.tika.TikaTextExtractor;
 import org.elasticsearch.client.Client;
+import org.junit.Assume;
+import org.junit.ClassRule;
+import org.junit.Test;
+import org.testcontainers.shaded.com.google.common.base.Throwables;
 
-public class ElasticSearchV1IntegrationTest extends  ElasticSearchIntegrationTest {
+public class ElasticSearchV1IntegrationTestWithoutTika extends  ElasticSearchIntegrationTest {
 
     private static final int SEARCH_SIZE = 1;
+
+
+    @ClassRule
+    public static TikaContainer tika = new TikaContainer();
+    private TikaTextExtractor textExtractor;
+
+    @Override
+    public void setUp() throws Exception {
+        textExtractor = new TikaTextExtractor(new TikaHttpClientImpl(TikaConfiguration.builder()
+            .host(tika.getIp())
+            .port(tika.getPort())
+            .timeoutInMillis(tika.getTimeoutInMillis())
+            .build()));
+        super.setUp();
+    }
+
+    @Override
+    protected TextExtractor getTextExtractor() {
+        return new DefaultTextExtractor();
+    }
 
     @Override
     protected MailboxMappingFactory provideMappingFactory() {
@@ -54,6 +84,12 @@ public class ElasticSearchV1IntegrationTest extends  ElasticSearchIntegrationTes
     @Override
     protected MessageToElasticSearchJson provideMessageToElasticSearchJson(TextExtractor textExtractor,
                                                                            ZoneId zoneId) {
-        return new MessageToElasticSearchJsonV1(textExtractor, zoneId, IndexAttachments.YES);
+        return new MessageToElasticSearchJsonV1(textExtractor, zoneId, IndexAttachments.NO);
+    }
+
+    @Override
+    @Test
+    public void searchWithPDFAttachmentShouldReturnMailsWhenAttachmentContentMatches() throws Exception {
+        Assume.assumeTrue(false);
     }
 }
